Updated the declaration of 'main' function.

+++ tests/code/valid/benchmark/fast/long/While_Valid_1.h
@@ -11,5 +11,5 @@ char** toString(long arr[], long size);
 void free_doublePtr(char** res, int res_size);
 void indirect_prlongf(char** res, int _res_size);
 long* reverse(long* _0, long _0_size);
-long main(long argc, char** argv);
+int main(int argc, char** argv);
